Update platform.h
This is perhaps an issue spread over all the sdk branches. The detection of little endianess only ever works for gcc, there's nothing for msvc. From my short research, it seems msvc doesn't/never tell the preprocessor anything about the CPU brand. So instead we're going to assume if the CPU is a x86 or x86_64 processor it's most likely intel so i386-family, therefore little endian. Even if there's room for error, this is still marginally better than not having anything defined.